Add `findall` method (#218)
Add `findall` method

BioSequences recently rewrote the entire set of `Base.find*` functions for v3, but we forgot `findall`.
This PR adds `Base.findall`. In particular, it adds the following two methods:

`Base.findall(pat, seq::BioSequence; overlap::Bool = DEFAULT_OVERLAP)`
`Base.findall(pat, seq::BioSequence, rng::UnitRange{Int}; overlap::Bool = DEFAULT_OVERLAP)`

Which mirrors the `find` functionality in Base, except defaulting the overlap to `true`, as we consider this
to be most useful.
Internally, this is implemented by first implementing `search`, which returns an internal `Search` object.
This can be iterated to yield each match, in order. `findall` simply calls `collect` on this iterator.
In this manner, we allow a lazy search which does not materialize all the results, while keeping in line
with the interface of `Base`.

This PR does not implement documentation - this will be added in a followup PR.

See also #190