Reset Julia compat to v1.5 (#245)
Commit 3c8f501 changed Julia compat to from 1.5 to 1.6, since the join method
used sum with init keyword, which was introduced in Julia 1.6.
However, BioSequences v3.0.0 was released with Julia compat 1.5. Upping Julia
compat in a patch release is not allowed. Hence, the fix from commit 3c8f501
is not applied in any released version.

This commit resets Julia compat to 1.5, and rephrases the join method so it no
longer uses any features not present in Julia 1.5. It has been tested with Julia
v 1.5 and passes.