Fix bug in zero_offset! called when materializing seqview (#261)
When materializing a seqview, a slice of the underlying buffer is copied
directly, for efficiency. Since the bytes may be placed in the buffer with a
nonzero offset, the optimised zero_offset! is called.
However, a bug in zero_offset! meant that it did not behave correctly when the
offset was already zero.