Error when translating gaps
DNA_Gap cannot be meaningfully translated, as it does not correspond to any
nucleotides, not even an unknown one. In fact, it's dubious that it's even a
nucleotide at all.
This PR makes `translate(!)` error when run on sequences with gaps, whereas
before this PR this was undefined behaviour (out-of-bounds access).
Alternative solutions could be to silently skip gaps, which is biologically
meaningful, but might lead to strange errors, or to insert AA_Gap, which can
only be done if the gaps come in groups of three corresponding to a whole
gap codon.
In the future we could change behaviour to skip gaps.

See issue 277