Rewrite counting of BioSequence (#316)
This is a large commit with multiple changes:

1. Remove bitparallel compiler
The previous bitparallel compiler was not great. It made use of metaprogramming,
which
    - Is hard to reason about
    - Does not play nice with Revise
    - Cannot be debugged using a debugger or similar tooling
    - Confuses linters and IDEs
    - May explode codegen, impacting compile times

This might have been necessary in earlier versions of Julia, but the modern
Julia compiler is so good at constant evaluation and inlining that this is
almost never a good idea.

2. Add internal chunk iterator types
This is an internal abstraction, an iterator that yields the encoded data chunks
of LongSequence or LongSubSeq. For subsequences, the yielded chunks are aligned,
such that the iterator returns the exact same elements as a chunk iterator
over the correponding LongSequence. These iterators might be generally useful in
future BioSequences work, and is something I've thought about for some time.

3. Deprecate the functions `n_ambigous`, `n_gaps` and `n_certain`
These should instead be accessed via their corresponding `count` methods, as
described in the documentation.
The rationale is that these functions provide little value as standalone
functions, as they should work exactly like `count` with specific predicates.
As we might want to add more in the future, this simply balloons the number
of exported symbols.

4. Deprecate some meaningless two-sequence counting methods
The method `n_ambiguous(::BioSequence ::BioSequence)`, and the equivalent for
n_gaps, n_certain and the `count` methods corresponding to these are strange
and obscure. Why are they there?
They significantly complicate the implementation of counting, while providing
no meaningful biological operations. At least I have no idea what the biological
significance of these methods are

5. Deprecate calling `matches` or `mismatches` with differing seq lengths
It is not clear what the correct thing to do here is, and it may be confusing
for users, who may expect a pairwise alignment or similar.
Instead of guessing what the user wants, force them to take a view of the longer
sequence if these differ in length.

6. Add new optimised methods for counting occurrences of symbol in sequence
E.g. `count(==(DNA_A), dna"TAG")` and a similar method for `isequal`.