Add more optimised find methods for BioSequences (#331)
This adds SIMD optimised methods for `findprev` and `findnext` (and therefore, by extension `findfirst` and `findlast`) for the predicates `!isgap`, `iscertain`, `!iscertain`, `isambiguous` and `!isambiguous` for the default alphabets for LongSequence and LongSubSeq

The purpose of these is to find e.g. regions of N's in a sequence, or find the first symbol in a long DNA sequence which is not a gap or an unambiguous symbol.

Some benchmarks for AA - nucleotide benchmarks are even better:
```julia
s = randaaseq(100_000)
@btime findfirst(isgap, s)
@btime findfirst(isambiguous, s)
@btime findfirst(!iscertain, s)
for i in eachindex(s)
   s[i] = AA_B
end
@btime findfirst(!isambiguous, s)
@btime findfirst(iscertain, s)
```
Before:
  1.738 μs (0 allocations: 0 bytes)
  39.195 μs (0 allocations: 0 bytes)
  39.195 μs (0 allocations: 0 bytes)
  39.205 μs (0 allocations: 0 bytes)
  39.205 μs (0 allocations: 0 bytes)

After:
  1.894 μs (0 allocations: 0 bytes)
  3.270 μs (0 allocations: 0 bytes)
  5.500 μs (0 allocations: 0 bytes)
  3.240 μs (0 allocations: 0 bytes)
  4.530 μs (0 allocations: 0 bytes)